+++ src/main/kotlin/com/wafflestudio/csereal/core/about/api/v1/AboutController.kt
@@ -1,6 +1,5 @@
 package com.wafflestudio.csereal.core.about.api.v1
 
-import com.wafflestudio.csereal.common.aop.AuthenticatedStaff
 import com.wafflestudio.csereal.common.enums.LanguageType
 import com.wafflestudio.csereal.core.about.api.req.*
 import com.wafflestudio.csereal.core.about.api.res.AboutSearchResBody
@@ -10,7 +9,6 @@ import jakarta.validation.Valid
 import jakarta.validation.constraints.Positive
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
-import org.springframework.web.multipart.MultipartFile
 
 @RequestMapping("/api/v1/about")
 @RestController("AboutControllerV1")
@@ -32,15 +30,6 @@ class AboutController(
         return ResponseEntity.ok(aboutService.readAbout(language, postType))
     }
 
-    @AuthenticatedStaff
-    @PutMapping("/{postType}")
-    fun updateAbout(
-        @PathVariable postType: String,
-        @RequestPart request: UpdateAboutReq,
-        @RequestPart newMainImage: MultipartFile?,
-        @RequestPart newAttachments: List<MultipartFile>?
-    ) = aboutService.updateAbout(postType, request, newMainImage, newAttachments)
-
     @Deprecated("Use V2 API")
     @GetMapping("/student-clubs")
     fun readAllClubs(
@@ -49,34 +38,12 @@ class AboutController(
         return ResponseEntity.ok(aboutService.readAllClubs(language))
     }
 
-    @AuthenticatedStaff
-    @PostMapping("/facilities")
-    fun createFacilities(@RequestPart request: CreateFacReq, @RequestPart mainImage: MultipartFile?) =
-        aboutService.createFacilities(request, mainImage)
-
-    @AuthenticatedStaff
-    @PutMapping("/facilities/{id}")
-    fun updateFacility(
-        @PathVariable id: Long,
-        @RequestPart request: CreateFacReq,
-        @RequestPart newMainImage: MultipartFile?
-    ) = aboutService.updateFacility(id, request, newMainImage)
-
-    @AuthenticatedStaff
-    @DeleteMapping("/facilities/{id}")
-    fun deleteFacility(@PathVariable id: Long) = aboutService.deleteFacility(id)
-
     @GetMapping("/facilities")
     fun readAllFacilities(
         @RequestParam(required = false, defaultValue = "ko") language: String
     ): ResponseEntity<List<AboutDto>> {
         return ResponseEntity.ok(aboutService.readAllFacilities(language))
     }
-
-    @AuthenticatedStaff
-    @PutMapping("/directions/{id}")
-    fun updateDirection(@PathVariable id: Long, @RequestBody request: UpdateDescriptionReq) =
-        aboutService.updateDirection(id, request)
 
     @GetMapping("/directions")
     fun readAllDirections(
@@ -85,30 +52,12 @@ class AboutController(
         return ResponseEntity.ok(aboutService.readAllDirections(language))
     }
 
-    @AuthenticatedStaff
-    @PutMapping("/future-careers")
-    fun updateFutureCareersPage(@RequestBody request: UpdateDescriptionReq) =
-        aboutService.updateFutureCareersPage(request)
-
     @GetMapping("/future-careers")
     fun readFutureCareers(
         @RequestParam(required = false, defaultValue = "ko") language: String
     ): ResponseEntity<FutureCareersPage> {
         return ResponseEntity.ok(aboutService.readFutureCareers(language))
     }
-
-    @AuthenticatedStaff
-    @PostMapping("/future-careers/company")
-    fun createCompany(@RequestBody request: CreateCompanyReq) = aboutService.createCompany(request)
-
-    @AuthenticatedStaff
-    @PutMapping("/future-careers/company/{id}")
-    fun updateCompany(@PathVariable id: Long, @RequestBody request: CreateCompanyReq) =
-        aboutService.updateCompany(id, request)
-
-    @AuthenticatedStaff
-    @DeleteMapping("/future-careers/company/{id}")
-    fun deleteCompany(@PathVariable id: Long) = aboutService.deleteCompany(id)
 
     @GetMapping("/search/top")
     fun searchTopAbout(
