+++ src/main/kotlin/com/wafflestudio/csereal/core/about/database/AboutRepository.kt
@@ -19,12 +19,6 @@ interface AboutRepository : JpaRepository<AboutEntity, Long>, AboutCustomReposit
         languageType: LanguageType,
         postType: AboutPostType
     ): AboutEntity
-
-    fun findByLanguageAndPostTypeAndNameContaining(
-        languageType: LanguageType,
-        postType: AboutPostType,
-        keyword: String
-    ): AboutEntity
 }
 
 interface AboutCustomRepository {
